refactor : RoomApi 수정
