+++ src/test/java/com/pser/hotel/domain/hotel/api/RoomApiTest.java
@@ -1,12 +1,16 @@
 package com.pser.hotel.domain.hotel.api;
 
-import static com.pser.hotel.domain.hotel.util.Utils.createUser;
+import static com.pser.hotel.domain.hotel.util.Utils.createImageUrl;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.pser.hotel.domain.hotel.application.RoomService;
+import com.pser.hotel.domain.hotel.dto.request.RoomRequest;
 import com.pser.hotel.domain.hotel.dto.request.RoomSearchRequest;
 import com.pser.hotel.domain.member.domain.User;
+import java.time.LocalTime;
+import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -18,6 +22,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.web.PageableHandlerMethodArgumentResolver;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 @ExtendWith(MockitoExtension.class)
@@ -44,6 +49,28 @@ class RoomApiTest {
     }
 
     @Test
+    @DisplayName("객실 리스트 조회 테스트")
+    public void roomList() throws Exception {
+        long hotelId = 1;
+        mockMvc.perform(get(String.format("/hotels/%d/rooms", hotelId))
+                        .param("page", String.valueOf(pageRequest.getOffset()))
+                        .param("size", String.valueOf(pageRequest.getPageSize())))
+                .andExpect(status().isOk())
+                .andDo(MockMvcResultHandlers.print());
+    }
+
+    @Test
+    @DisplayName("단건 객실 조회 테스트")
+    public void roomDetails() throws Exception {
+        long hotelId = 1;
+        long roomId = 1;
+        String url = String.format("/hotels/%d/rooms/%d", hotelId, roomId);
+        mockMvc.perform(get(url))
+                .andExpect(status().isOk())
+                .andDo(MockMvcResultHandlers.print());
+    }
+
+    @Test
     @DisplayName("객실 삭제 테스트")
     public void roomDelete() throws Exception {
         // Given
@@ -62,5 +89,39 @@ class RoomApiTest {
     public void roomDeleteByBadRequest() throws Exception {
         mockMvc.perform(delete("/hotels/1/rooms/1"))
                 .andExpect(status().isBadRequest());
+    }
+
+    private RoomRequest createRoomRequestDto() {
+        return new RoomRequest(
+                "객실이름", "설명", "mainImage.url", "주의사항", 1000, LocalTime.of(15, 00), LocalTime.of(11, 00),
+                1, 1, 1,
+                true, true, true, true,
+                true, true, true, true,
+                true, true, true, true, true, true,
+                List.of(createImageUrl(), createImageUrl())
+        );
+    }
+
+    private User createUser() {
+        return User.builder().email("test_email").password("123").build();
+    }
+
+    public String getRoomByJson() {
+        RoomRequest dto = createRoomRequestDto();
+        return String.format(
+                "{\"name\" : \"%s\", \"description\" : \"%s\","
+                        + " \"precaution\" : \"%s\", \"price\" : \"%d\", \"standardCapacity\" : \"%d\","
+                        + " \"maxCapacity\" : \"%d\", \"totalRooms\" : \"%d\", \"heatingSystem\" : \"%s\","
+                        + " \"tv\" : \"%s\", \"refrigerator\" : \"%s\", \"airConditioner\" : \"%s\","
+                        + " \"washer\" : \"%s\", \"terrace\" : \"%s\", \"coffeeMachine\" : \"%s\","
+                        + " \"internet\" : \"%s\", \"kitchen\" : \"%s\", \"bathtub\" : \"%s\","
+                        + " \"iron\" : \"%s\", \"pool\" : \"%s\", \"pet\" : \"%s\","
+                        + " \"inAnnex\" : \"%s\"}"
+                , dto.getName(), dto.getDescription(), dto.getPrecaution(),
+                dto.getPrice(), dto.getStandardCapacity(), dto.getMaxCapacity(), dto.getTotalRooms(),
+                dto.getHeatingSystem(), dto.getTv(), dto.getRefrigerator(), dto.getAirConditioner(),
+                dto.getWasher(), dto.getTerrace(),
+                dto.getCoffeeMachine(), dto.getInternet(), dto.getKitchen(), dto.getBathtub(), dto.getIron(),
+                dto.getPool(), dto.getPet(), dto.getInAnnex());
     }
 }
